refactor: target to esnext, move slackData to lib so members.ts can use it for up to date member count, and lib to ts progress
