--- lib/fetcher.js
@@ -1,9 +1,0 @@
-/**
- * useSWR() fetcher
- * @param  {...any} args fetcher for useSWR()
- */
-
-export default async function fetcher(...args) {
-  const res = await fetch(...args)
-  return await res.json()
-}
