+++ lib/fetcher.ts
@@ -1,0 +1,7 @@
+/**
+ * useSWR() fetcher
+ */
+export default async function fetcher(...args: Parameters<typeof fetch>) {
+  const res = await fetch(...args)
+  return await res.json()
+}
