--- lib/members.js
@@ -1,3 +1,0 @@
-export const count = 87752
-export const formatted = count.toLocaleString('en-US')
-export const thousands = Math.round(count / 1000)
